requirements.txt
# Core
torch>=1.10.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.0.0
plotly>=5.0.0

# Configuration (optional)
omegaconf>=2.1.0
hydra-core>=1.1.0
PyYAML>=5.4.0

# StarCraft II Environment (optional, for data collection)
# pysc2>=3.0.0
# absl-py>=0.13.0

# Development
tqdm>=4.62.0

# Live Game Bridge (web → SC2)
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.28.0
sse-starlette>=1.6.0
